#331 Move clockA12 into the VT02 test harness
